requirements.txt
Django==2.0.4
django-storages==1.6.6
djangorestframework==3.8.1
gunicorn==19.7.1
Pillow==5.1.0
psycopg2==2.7.4
pytz==2018.3
